Add localstorage for global state
