+++ .gitignore
@@ -30,3 +30,4 @@ gradlew
 gradlew.bat
 gradle/
 !gradle-wrapper.jar
+.*iml
